Move service_extension_widgets.dart to service/ dir
