+++ packages/devtools_app/lib/src/screens/performance/performance_screen.dart
@@ -12,6 +12,7 @@ import '../../analytics/analytics_common.dart';
 import '../../analytics/constants.dart' as analytics_constants;
 import '../../config_specific/import_export/import_export.dart';
 import '../../primitives/auto_dispose_mixin.dart';
+import '../../service/service_extension_widgets.dart';
 import '../../service/service_extensions.dart' as extensions;
 import '../../shared/banner_messages.dart';
 import '../../shared/common_widgets.dart';
@@ -23,7 +24,6 @@ import '../../shared/split.dart';
 import '../../shared/theme.dart';
 import '../../shared/version.dart';
 import '../../ui/icons.dart';
-import '../../ui/service_extension_widgets.dart';
 import '../../ui/vm_flag_widgets.dart';
 import 'enhance_tracing.dart';
 import 'event_details.dart';
